### 2021-11-04 #3891 @whedon
[PRE REVIEW]: Council Data Project: Software for Municipal Data Collection, Analysis, and Publication
**Submitting author:** @JacksonMaxfield (<a href="http://orcid.org/0000-0003-2564-0373">Jackson Brown</a>)
**Repository:** <a href="https://github.com/CouncilDataProject/cookiecutter-cdp-deployment" target ="_blank">https://github.com/CouncilDataProject/cookiecutter-cdp-deployment</a>
**Version:** v3.0.0
