### 2021-11-04 #3891 @JacksonMaxfield
Hey @danielskatz!

Thanks @nniiicc for pointing to the issue. The software which should be reviewed is:

* https://github.com/CouncilDataProject/cookiecutter-cdp-deployment (the repository which this paper is in)
* https://github.com/CouncilDataProject/cdp-backend
* https://github.com/CouncilDataProject/cdp-frontend
