### 2021-11-05 #3891 @danielskatz
Thanks @JacksonMaxfield and @nniiicc - I've ended the scope query, and we'll move onto finding an editor
