+++ paper/paper.bib
@@ -1,15 +1,14 @@
-@article{Trounstine:2009,
-  issn      = {15375927, 15410986},
-  url       = {http://www.jstor.org/stable/40407007},
-  abstract  = {The study of local politics has been relegated to the periphery of political science and many explanations have been offered for the marginalization of the subfield. I offer three related arguments for why scholars should revisit the study of sub-state politics. First, the local level is the source of numerous political outcomes that matter because they represent a large proportion of political events in the United States. Secondly, there are methodological advantages to studying local politics. Finally, analyzing politics at the sub-state level can generate thoroughly different kinds of questions than a purely national-level focus and can offer different answers to questions that apply more generally. Research on local politics can and should contribute to broader debates in political science and ensure that we understand both how and why cities are unique.},
-  author    = {Jessica Trounstine},
-  journal   = {Perspectives on Politics},
-  number    = {3},
-  pages     = {611--618},
-  publisher = {[American Political Science Association, Cambridge University Press]},
+@article{Trounstine2009,
   title     = {All Politics Is Local: The Reemergence of the Study of City Politics},
   volume    = {7},
-  year      = {2009}
+  doi       = {10.1017/S1537592709990892},
+  url       = {https://doi.org/10.1017/S1537592709990892},
+  number    = {3},
+  journal   = {Perspectives on Politics},
+  publisher = {Cambridge University Press},
+  author    = {Trounstine, Jessica},
+  year      = {2009},
+  pages     = {611–618}
 }
 
 @misc{cookiecutter,
