### 2021-11-09 #3891 @JacksonMaxfield
DOI for "All Politics is Local" updated but CourtListener doesn't seem to have a DOI at all so leaving as is.
