+++ SETUP/Makefile
@@ -22,24 +22,24 @@ login: ## Run GCloud and Pulumi logins
 	gcloud auth application-default login
 
 init: ## Create google project and pulumi stack
-	gcloud projects create cdp-example-fsijkyuz --set-as-default
-	pulumi stack init cdp-example-fsijkyuz
+	gcloud projects create cdp-example-usptvjgo --set-as-default
+	pulumi stack init cdp-example-usptvjgo
 
 set-cors: ## Set cors for the bucket
-	gsutil cors set cors.json gs://cdp-example-fsijkyuz.appspot.com/
+	gsutil cors set cors.json gs://cdp-example-usptvjgo.appspot.com/
 
 destroy: ## Fully teardown Pulumi and GCloud
-	pulumi stack rm cdp-example-fsijkyuz --force
-	gcloud projects delete cdp-example-fsijkyuz
-	rm -f .keys/cdp-example-fsijkyuz.json
+	pulumi stack rm cdp-example-usptvjgo --force
+	gcloud projects delete cdp-example-usptvjgo
+	rm -f .keys/cdp-example-usptvjgo.json
 
 gen-key: ## Generate a service account JSON
 	mkdir .keys/ -p
-	gcloud iam service-accounts create cdp-example-fsijkyuz \
+	gcloud iam service-accounts create cdp-example-usptvjgo \
     	--description="CDP Dev Service Account for $(USER)" \
-    	--display-name="cdp-example-fsijkyuz"
-	gcloud projects add-iam-policy-binding cdp-example-fsijkyuz \
-    	--member="serviceAccount:cdp-example-fsijkyuz@cdp-example-fsijkyuz.iam.gserviceaccount.com" \
+    	--display-name="cdp-example-usptvjgo"
+	gcloud projects add-iam-policy-binding cdp-example-usptvjgo \
+    	--member="serviceAccount:cdp-example-usptvjgo@cdp-example-usptvjgo.iam.gserviceaccount.com" \
 		--role="roles/owner"
-	gcloud iam service-accounts keys create .keys/cdp-example-fsijkyuz.json \
-  		--iam-account "cdp-example-fsijkyuz@cdp-example-fsijkyuz.iam.gserviceaccount.com"
+	gcloud iam service-accounts keys create .keys/cdp-example-usptvjgo.json \
+  		--iam-account "cdp-example-usptvjgo@cdp-example-usptvjgo.iam.gserviceaccount.com"
