+++ infra/__main__.py
@@ -7,7 +7,7 @@ from pulumi import export
 ###############################################################################
 
 cdp_stack = CDPStack(
-    gcp_project_id="cdp-example-usptvjgo",
+    gcp_project_id="cdp-example-bubumbnu",
     firestore_location="us-central",
 )
 
