+++ SETUP/README.md
@@ -48,16 +48,16 @@ There are additional tasks required after generating this repository.
     ```
 
 1.  Create (or re-use) a
-    [Google Cloud billing account](https://console.cloud.google.com/billing/linkedaccount?project=cdp-example-usptvjgo)
-    and attach it to the newly created project (cdp-example-usptvjgo).
+    [Google Cloud billing account](https://console.cloud.google.com/billing/linkedaccount?project=cdp-example-bubumbnu)
+    and attach it to the newly created project (cdp-example-bubumbnu).
 
     For more details on the cost of maintaining a CDP Instance, see our [estimated cost breakdown](https://github.com/CouncilDataProject/cookiecutter-cdp-deployment#cost).
 
 1.  Generate a Google Service Account JSON Key for your Google Cloud Project.
 
     This will create a directory called `.keys` within this `SETUP` directory and
-    add a file called `cdp-example-usptvjgo.json` to it
-    (i.e. `.keys/cdp-example-usptvjgo)`. This file will be used later on.
+    add a file called `cdp-example-bubumbnu.json` to it
+    (i.e. `.keys/cdp-example-bubumbnu)`. This file will be used later on.
 
     Run:
 
@@ -77,7 +77,7 @@ There are additional tasks required after generating this repository.
     2. Google Service Account JSON -- Create a [new secret](https://github.com/CouncilDataProject/example/settings/secrets/actions/new)
 
     -   Set the name to: **GOOGLE_CREDENTIALS**
-    -   Set the value to: the contents of the file `.keys/cdp-example-usptvjgo.json`
+    -   Set the value to: the contents of the file `.keys/cdp-example-bubumbnu.json`
     -   Click "Add secret"
 
 1.  Initialize and push the local repository to GitHub.
@@ -140,7 +140,7 @@ There are additional tasks required after generating this repository.
     configure Firebase Security Rules.
 
     -   Navigate to [Firebase Console](https://console.firebase.google.com),
-        login to the Google Account you used during step #2, select the `cdp-example-usptvjgo` Firebase project
+        login to the Google Account you used during step #2, select the `cdp-example-bubumbnu` Firebase project
         -   Navigate to "Firestore Database", select the "Rules" tab, paste the following in:
             ```
             rules_version = '2';
