+++ SETUP/Makefile
@@ -22,24 +22,24 @@ login: ## Run GCloud and Pulumi logins
 	gcloud auth application-default login
 
 init: ## Create google project and pulumi stack
-	gcloud projects create cdp-example-usptvjgo --set-as-default
-	pulumi stack init cdp-example-usptvjgo
+	gcloud projects create cdp-example-bubumbnu --set-as-default
+	pulumi stack init cdp-example-bubumbnu
 
 set-cors: ## Set cors for the bucket
-	gsutil cors set cors.json gs://cdp-example-usptvjgo.appspot.com/
+	gsutil cors set cors.json gs://cdp-example-bubumbnu.appspot.com/
 
 destroy: ## Fully teardown Pulumi and GCloud
-	pulumi stack rm cdp-example-usptvjgo --force
-	gcloud projects delete cdp-example-usptvjgo
-	rm -f .keys/cdp-example-usptvjgo.json
+	pulumi stack rm cdp-example-bubumbnu --force
+	gcloud projects delete cdp-example-bubumbnu
+	rm -f .keys/cdp-example-bubumbnu.json
 
 gen-key: ## Generate a service account JSON
 	mkdir .keys/ -p
-	gcloud iam service-accounts create cdp-example-usptvjgo \
+	gcloud iam service-accounts create cdp-example-bubumbnu \
     	--description="CDP Dev Service Account for $(USER)" \
-    	--display-name="cdp-example-usptvjgo"
-	gcloud projects add-iam-policy-binding cdp-example-usptvjgo \
-    	--member="serviceAccount:cdp-example-usptvjgo@cdp-example-usptvjgo.iam.gserviceaccount.com" \
+    	--display-name="cdp-example-bubumbnu"
+	gcloud projects add-iam-policy-binding cdp-example-bubumbnu \
+    	--member="serviceAccount:cdp-example-bubumbnu@cdp-example-bubumbnu.iam.gserviceaccount.com" \
 		--role="roles/owner"
-	gcloud iam service-accounts keys create .keys/cdp-example-usptvjgo.json \
-  		--iam-account "cdp-example-usptvjgo@cdp-example-usptvjgo.iam.gserviceaccount.com"
+	gcloud iam service-accounts keys create .keys/cdp-example-bubumbnu.json \
+  		--iam-account "cdp-example-bubumbnu@cdp-example-bubumbnu.iam.gserviceaccount.com"
