### 2021-11-11 #3904 @JacksonMaxfield
Excited to see this move forward! Thanks all for agreeing to review.

As a reminder this is three tools together:
* https://github.com/CouncilDataProject/cookiecutter-cdp-deployment (the repo the paper is in)
* https://github.com/CouncilDataProject/cdp-backend
* https://github.com/CouncilDataProject/cdp-frontend
