+++ README.md
@@ -76,6 +76,7 @@ For more details on what is created from using this cookiecutter template, see [
 | Parameter                      | Description                                                                                                                 | Example 1                                      | Example 2                                         |
 | ------------------------------ | --------------------------------------------------------------------------------------------------------------------------- | ---------------------------------------------- | ------------------------------------------------- |
 | municipality                   | The name of the municipality (town, city, county, etc.) that this CDP Instance will store data for.                         | Seattle                                        | King County                                       |
+| governing_body_type            | What type of governing body this instance is for.                                                                           | city council                                   | county council                                    |
 | municipality_slug              | The name of the municipality cleaned for use in the web application and parts of repository naming.                         | seattle                                        | king-county                                       |
 | python_municipality_slug       | The name of the municipality cleaned for use in specifically Python parts of the application.                               | seattle                                        | king_county                                       |
 | infrastructure_slug            | The name of the municipality cleaned for use in specifically application infrastructure. Must be globally unique to GCP.    | cdp-seattle-abasjkqy                           | cdp-king-county-uiqmsbaw                          |
