+++ .github/ISSUE_TEMPLATE/_new_instance_form.yml
@@ -24,6 +24,14 @@ body:
     validations:
       required: true
   - type: input
+    id: governing_body_type
+    attributes:
+      label: Governing Body Type
+      description: What type of governing body this instance is for. Must be one of, 'city_council', 'county_council', 'other', or 'school_board'.
+      placeholder: ex. "city council", "county council"
+    validations:
+      required: true
+  - type: input
     id: maintainer_github_name
     attributes:
       label: Maintainer GitHub Name
