+++ {{ cookiecutter.hosting_github_repo_name }}/infra/__main__.py
@@ -12,7 +12,7 @@ cdp_stack = CDPStack(
     firestore_location="{{ cookiecutter.firestore_region }}",
     hosting_github_url="{{ cookiecutter.hosting_github_url }}",
     hosting_web_app_address="{{ cookiecutter.hosting_web_app_address }}",
-    governing_body="{{ cookiecutter.governing_body_type }}"
+    governing_body="{{ cookiecutter.governing_body_type }}",
 )
 
 export("firestore_address", cdp_stack.firestore_app.app_id)
