+++ cookiecutter.yaml
@@ -1,8 +1,9 @@
 default_context:
     municipality: "Example"
+    governing_body_type: "city council"
     municipality_slug: "example"
     python_municipality_slug: "example"
-    infrastructure_slug: "cdp-example-bubumbnu"
+    infrastructure_slug: "cdp-example-qnghgagv"
     maintainer_or_org_full_name: "Council Data Project Contributors"
     hosting_github_username_or_org: "CouncilDataProject"
     hosting_github_repo_name: "example"
