+++ SETUP/Makefile
@@ -22,24 +22,24 @@ login: ## Run GCloud and Pulumi logins
 	gcloud auth application-default login
 
 init: ## Create google project and pulumi stack
-	gcloud projects create cdp-example-qnghgagv --set-as-default
-	pulumi stack init cdp-example-qnghgagv
+	gcloud projects create cdp-example-xgszbeub --set-as-default
+	pulumi stack init cdp-example-xgszbeub
 
 set-cors: ## Set cors for the bucket
-	gsutil cors set cors.json gs://cdp-example-qnghgagv.appspot.com/
+	gsutil cors set cors.json gs://cdp-example-xgszbeub.appspot.com/
 
 destroy: ## Fully teardown Pulumi and GCloud
-	pulumi stack rm cdp-example-qnghgagv --force
-	gcloud projects delete cdp-example-qnghgagv
-	rm -f .keys/cdp-example-qnghgagv.json
+	pulumi stack rm cdp-example-xgszbeub --force
+	gcloud projects delete cdp-example-xgszbeub
+	rm -f .keys/cdp-example-xgszbeub.json
 
 gen-key: ## Generate a service account JSON
 	mkdir .keys/ -p
-	gcloud iam service-accounts create cdp-example-qnghgagv \
+	gcloud iam service-accounts create cdp-example-xgszbeub \
     	--description="CDP Dev Service Account for $(USER)" \
-    	--display-name="cdp-example-qnghgagv"
-	gcloud projects add-iam-policy-binding cdp-example-qnghgagv \
-    	--member="serviceAccount:cdp-example-qnghgagv@cdp-example-qnghgagv.iam.gserviceaccount.com" \
+    	--display-name="cdp-example-xgszbeub"
+	gcloud projects add-iam-policy-binding cdp-example-xgszbeub \
+    	--member="serviceAccount:cdp-example-xgszbeub@cdp-example-xgszbeub.iam.gserviceaccount.com" \
 		--role="roles/owner"
-	gcloud iam service-accounts keys create .keys/cdp-example-qnghgagv.json \
-  		--iam-account "cdp-example-qnghgagv@cdp-example-qnghgagv.iam.gserviceaccount.com"
+	gcloud iam service-accounts keys create .keys/cdp-example-xgszbeub.json \
+  		--iam-account "cdp-example-xgszbeub@cdp-example-xgszbeub.iam.gserviceaccount.com"
