+++ infra/__main__.py
@@ -7,7 +7,7 @@ from pulumi import export
 ###############################################################################
 
 cdp_stack = CDPStack(
-    gcp_project_id="cdp-example-xgszbeub",
+    gcp_project_id="cdp-example-ykfomjki",
     municipality_name="Example",
     firestore_location="us-central",
     hosting_github_url="https://github.com/CouncilDataProject/example",
