+++ SETUP/Makefile
@@ -22,24 +22,24 @@ login: ## Run GCloud and Pulumi logins
 	gcloud auth application-default login
 
 init: ## Create google project and pulumi stack
-	gcloud projects create cdp-example-xgszbeub --set-as-default
-	pulumi stack init cdp-example-xgszbeub
+	gcloud projects create cdp-example-ykfomjki --set-as-default
+	pulumi stack init cdp-example-ykfomjki
 
 set-cors: ## Set cors for the bucket
-	gsutil cors set cors.json gs://cdp-example-xgszbeub.appspot.com/
+	gsutil cors set cors.json gs://cdp-example-ykfomjki.appspot.com/
 
 destroy: ## Fully teardown Pulumi and GCloud
-	pulumi stack rm cdp-example-xgszbeub --force
-	gcloud projects delete cdp-example-xgszbeub
-	rm -f .keys/cdp-example-xgszbeub.json
+	pulumi stack rm cdp-example-ykfomjki --force
+	gcloud projects delete cdp-example-ykfomjki
+	rm -f .keys/cdp-example-ykfomjki.json
 
 gen-key: ## Generate a service account JSON
 	mkdir .keys/ -p
-	gcloud iam service-accounts create cdp-example-xgszbeub \
+	gcloud iam service-accounts create cdp-example-ykfomjki \
     	--description="CDP Dev Service Account for $(USER)" \
-    	--display-name="cdp-example-xgszbeub"
-	gcloud projects add-iam-policy-binding cdp-example-xgszbeub \
-    	--member="serviceAccount:cdp-example-xgszbeub@cdp-example-xgszbeub.iam.gserviceaccount.com" \
+    	--display-name="cdp-example-ykfomjki"
+	gcloud projects add-iam-policy-binding cdp-example-ykfomjki \
+    	--member="serviceAccount:cdp-example-ykfomjki@cdp-example-ykfomjki.iam.gserviceaccount.com" \
 		--role="roles/owner"
-	gcloud iam service-accounts keys create .keys/cdp-example-xgszbeub.json \
-  		--iam-account "cdp-example-xgszbeub@cdp-example-xgszbeub.iam.gserviceaccount.com"
+	gcloud iam service-accounts keys create .keys/cdp-example-ykfomjki.json \
+  		--iam-account "cdp-example-ykfomjki@cdp-example-ykfomjki.iam.gserviceaccount.com"
