+++ .github/ISSUE_TEMPLATE/_new_instance_form.yml
@@ -49,6 +49,9 @@ body:
         To find out if your municipality uses Legistar for its legislation management software follow the steps outlined in our [Legistar Documentation](https://councildataproject.org/cdp-scrapers/finding_legistar_id.html). If Legistar is used, this instance configuration process could potentially be automated for you and you can fill in your Legistar Id below. If not, you will need to write a [custom event scraper](https://councildataproject.org/cdp-scrapers/index.html#creating-a-custom-scraper) to complete the process.
         
         If you plan on using Legistar, all fields in this section are required. I.e. the municipality's Legistar Client Id is required and can be found by following our recommendations in [documentation](https://councildataproject.org/cdp-scrapers/legistar_scraper.html) and the timezone for your municipality is required.
+
+        **Note:** If you already have a scraper written and published in `cdp-scrapers`,
+        you can provide the municipality slug instead of these Legistar options.
   - type: input
     id: legistar_client_id
     attributes:
