### 2021-11-20 #3904 @chainsawriot
@JacksonMaxfield @chartgerink

I read with interest the paper by @JacksonMaxfield et al. I think the entire software suite is useful for political scientists, among others, who are interested in US local politics. I am very impressed with the [live demo](https://councildataproject.org/seattle/#/).

The software uses speech-to-text technology from Google to make speeches machine-readable and searchable in a Firebase database. The software is well-documented. I've attempted to create a template. Due to the cost, I've not tried to deploy the project.

I only have two minor concerns related to the project and the paper. Other than the two, I don't have further concerns.

# Project

* The project should have clear [community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines).

# Paper

* The text labels in Figure 1 are not easy to read. The main idea in the figure is those labels, not the icons.
