### 2021-11-21 #3904 @hknd23
Hi @JacksonMaxfield,

The CDP project has many potential applications in the social sciences such as Political Science, Public Policy, and Sociology, as well as Legal Studies. The website and interface are well developed, and the ability to collect rich data is noteworthy.
Please note that due to the cost to host my own instance, I also have not deployed the project myself. However, I have followed the steps up to the creation of the repository to host the instance with few suggestions.
There are a few remaining points that can help the project to be more user-friendly:

1. Installation instructions:
a. As the self-deployment of the instances requires the use of gcloud, gsutil, and Pulumi. In addition to the things to know section, it may be helpful to first-time users to mention installing these tools early on in the installation instructions.
b. The step-by-step Youtube video is extremely helpful as a tutorial detailing the steps needed to deploy an instance. In addition to this, you should also have a tutorial page as a walkthrough guide and example deployed repository. These can serve as quick reference tools and are easier to update/maintain as you add more features in the future.

2. Community guidelines: There are CONTRIBUTING files for the frontend and backend repositories. The main repositories should similarly have contributing guidelines and link to the guidelines in the other two as well.

3. The text in the figure is hard to read. I agree with @chainsawriot that the icons are not informative as their labels. Breaking the figure down to smaller steps or a list describing the process may help.

Overall, this is a very interesting project that can facilitate research progress in local governments. Please let me know if you have questions about my comments.
