### 2021-11-22 #3904 @chartgerink
Thank you @chainsawriot and @hknd23 for the constructive reviews! 🥳

@JacksonMaxfield the ball is back in your court regarding the reviewer's comments. I agree that the main points to respond to are:

* Installation instructions
* Figure 1 readability
* Community guidelines

Please feel free to ping me here once you feel you've worked on these points to resolve the reviewer's questions/comments to your satisfaction.
