### 2021-11-22 #3904 @JacksonMaxfield
Thank you all!

I will work on the community guidelines + contributing for the cookiecutter repository and I will work on the figure 1 readability.

A follow up question for @hknd23 regarding installation steps:

> As the self-deployment of the instances requires the use of gcloud, gsutil, and Pulumim, in addition to the things to know section, it may be helpful to first-time users to mention installing these tools early on in the installation instructions.

Agree. Will add a "prerequisites" section.

> In addition to this, you should also have a tutorial page as a walkthrough guide and example deployed repository. These can serve as quick reference tools and are easier to update/maintain as you add more features in the future.

These [SETUP steps](https://github.com/CouncilDataProject/cookiecutter-cdp-deployment/tree/main/%7B%7B%20cookiecutter.hosting_github_repo_name%20%7D%7D/SETUP) are provided with the generated repository. Can you clarify what should be added or should those directions simply be copied to the README or something similar? An entirely new documentation page?
