### 2021-11-23 #3904 @hknd23
Hi @JacksonMaxfield,

The instructions from SETUP look great. I was able to follow the steps without much issue when I first set up my example, and my impression is that these steps are the same with every generated repository. My suggestion is to have a section for or link to these steps in the main installation instructions so users can see them before having to set up the generated repository. I hope this answers your question. Please let me know if you have other questions.
