### 2021-11-23 a9f62d6
Add details about SETUP steps to base README
