### 2021-11-24 46a2457
Update core infra image
