+++ paper/assets/requirements.txt
@@ -1,1 +1,0 @@
-diagrams>=0.20,<1
