+++ README.md
@@ -61,10 +61,10 @@ you will need to install `cookiecutter` and use this template.
 
 Before you begin, please note that you will need to install or have available the following:
 
-* [gcloud](https://cloud.google.com/sdk/docs/install)
-* [pulumi](https://www.pulumi.com/docs/get-started/install/)
-* [gsutil](https://cloud.google.com/storage/docs/gsutil_install)
-* [Python 3.6+](https://www.python.org/downloads/) (Any Python version greater than or equal to 3.6)
+-   [gcloud](https://cloud.google.com/sdk/docs/install)
+-   [pulumi](https://www.pulumi.com/docs/get-started/install/)
+-   [gsutil](https://cloud.google.com/storage/docs/gsutil_install)
+-   [Python 3.6+](https://www.python.org/downloads/) (Any Python version greater than or equal to 3.6)
 
 Once all tools are installed, the rest of the infrastructure setup process
 should take about 15 minutes.
@@ -86,19 +86,19 @@ For more details on what is created from using this cookiecutter template, see [
 
 The short summary of setup tasks remaining are:
 
-* The creation of a new GitHub repository for the instance.
-* Logging in or creating accounts for Google Cloud and Pulumi.
-* Initialize the basic infrastructure.
-* Assign a billing account to the created Google Cloud project.
-* Generate credentials for the Google Project for use in automated scripts.
-* Attach credentials as secrets to the GitHub repository.
-* Push the cookiecutter generated files to the GitHub repository.
-* Setup web hosting through GitHub Pages.
-* Enable open access for data with CORS rules and public read.
-* Write an event scraper for your municipality (it may be useful to
-  build off of [cdp-scrapers](https://github.com/CouncilDataProject/cdp-scrapers))
+-   The creation of a new GitHub repository for the instance.
+-   Logging in or creating accounts for Google Cloud and Pulumi.
+-   Initialize the basic infrastructure.
+-   Assign a billing account to the created Google Cloud project.
+-   Generate credentials for the Google Project for use in automated scripts.
+-   Attach credentials as secrets to the GitHub repository.
+-   Push the cookiecutter generated files to the GitHub repository.
+-   Setup web hosting through GitHub Pages.
+-   Enable open access for data stored by Google Cloud and Firebase.
+-   Write an event scraper for your municipality (it may be useful to
+    build off of [cdp-scrapers](https://github.com/CouncilDataProject/cdp-scrapers))
 
-You can also an example generated repository and the full steps listed
+You can also see an example generated repository and the full steps listed
 [here](https://github.com/CouncilDataProject/cookiecutter-cdp-deployment/tree/example-build/SETUP).
 
 ### Cookiecutter Parameters
