### 2021-11-24 d5014f5
Resolve Hawks comments
