### 2021-11-24 a3c2028
admin/resolve-joss-reviewer-comments (#67)
* Add CONTRIBUTING

* Add details about SETUP steps to base README

* Add link to example build SETUP

* Update core infra image

* Downsample core infra image and remove assets README

* Resolve Isaac's comments

* Resolve Hawks comments
