+++ paper/assets/README.md
@@ -1,10 +1,0 @@
-# Figure Generation
-
-## Setup
-
-1. `pip install -r requirements.txt`
-2. [Install graphviz](https://graphviz.org/download/)
-
-## Infrastructure
-
-Run: `python infra.py`
