### 2021-11-24 #3904 @JacksonMaxfield
Hello all! (cc @chartgerink)

I just pushed some changes that I believe resolve your comments: https://github.com/CouncilDataProject/cookiecutter-cdp-deployment/commit/a3c202858a1cfa9d6c4a2619dffb0188410fbeb0

Thanks again for your comments :slightly_smiling_face:
