### 2021-11-24 #3904 @chainsawriot
@JacksonMaxfield Thanks for updating the paper. I have checked the community guidelines box above.

I don't want to be a "Reviewer 2". The figure actually looks [okay on its own](https://github.com/chainsawriot/cookiecutter-cdp-deployment/blob/main/paper/assets/cdp_core_infrastructure.png). However, it got resized in the paper because it needs to be inside that small column. Sorry for my old eyes, but I need to magnify 300% to read the text. For the graph to be readable, I believe the text should have the same size as the caption in the paper.

I am trying to redo the figure for you. I will submit a PR later.
