### 2021-11-24 #3904 @chainsawriot
@JacksonMaxfield I am halfway creating it. I have tried to test it by embedding it in your paper. It works great with this size.

https://docs.google.com/drawings/d/1IPebeFrgqOLVRpuANxOXkxZ7I54DYC3EOKWFYbcdZck/edit?usp=sharing
