### 2021-11-28 58b1364
admin/explicit-version-pin-for-dependabot  (#70)
* Explicit versioning for cdp-backend and cdp-frontend

* Pin cdp-backend version in infra requirements.txt
