### 2021-11-29 #3904 @chartgerink
Thank you all for such an engaged review process! Fantastic to see 😊

@hknd23 there are two open checks on your checklist - are these resolved to your satisfaction based on the author's response? If so, could you mark those as done as well?

@chainsawriot thanks for being so kind to help out with the figure! @JacksonMaxfield feel free to regenerate the PDF when that's ready (you seem to know your way around Whedon 😉 )
