### 2021-11-29 8d9180b
admin/update-acknowledgements-in-joss-paper (#68)
* Update acknowledgments

* Update acknowledgements and fig 1 description

* Update fig 1
