+++ SETUP/Makefile
@@ -22,24 +22,24 @@ login: ## Run GCloud and Pulumi logins
 	gcloud auth application-default login
 
 init: ## Create google project and pulumi stack
-	gcloud projects create cdp-example-nvkadfyj --set-as-default
-	pulumi stack init cdp-example-nvkadfyj
+	gcloud projects create cdp-example-wquqtqng --set-as-default
+	pulumi stack init cdp-example-wquqtqng
 
 set-cors: ## Set cors for the bucket
-	gsutil cors set cors.json gs://cdp-example-nvkadfyj.appspot.com/
+	gsutil cors set cors.json gs://cdp-example-wquqtqng.appspot.com/
 
 destroy: ## Fully teardown Pulumi and GCloud
-	pulumi stack rm cdp-example-nvkadfyj --force
-	gcloud projects delete cdp-example-nvkadfyj
-	rm -f .keys/cdp-example-nvkadfyj.json
+	pulumi stack rm cdp-example-wquqtqng --force
+	gcloud projects delete cdp-example-wquqtqng
+	rm -f .keys/cdp-example-wquqtqng.json
 
 gen-key: ## Generate a service account JSON
 	mkdir .keys/ -p
-	gcloud iam service-accounts create cdp-example-nvkadfyj \
+	gcloud iam service-accounts create cdp-example-wquqtqng \
     	--description="CDP Dev Service Account for $(USER)" \
-    	--display-name="cdp-example-nvkadfyj"
-	gcloud projects add-iam-policy-binding cdp-example-nvkadfyj \
-    	--member="serviceAccount:cdp-example-nvkadfyj@cdp-example-nvkadfyj.iam.gserviceaccount.com" \
+    	--display-name="cdp-example-wquqtqng"
+	gcloud projects add-iam-policy-binding cdp-example-wquqtqng \
+    	--member="serviceAccount:cdp-example-wquqtqng@cdp-example-wquqtqng.iam.gserviceaccount.com" \
 		--role="roles/owner"
-	gcloud iam service-accounts keys create .keys/cdp-example-nvkadfyj.json \
-  		--iam-account "cdp-example-nvkadfyj@cdp-example-nvkadfyj.iam.gserviceaccount.com"
+	gcloud iam service-accounts keys create .keys/cdp-example-wquqtqng.json \
+  		--iam-account "cdp-example-wquqtqng@cdp-example-wquqtqng.iam.gserviceaccount.com"
