### 2021-11-30 #3904 @chartgerink
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1017/S1537592709990892 is OK
> - 10.21105/joss.00411 is OK
>
> MISSING DOIs
>
> - 10.5860/choice.49-1821 may be a valid DOI for title: CourtListener
>
> INVALID DOIs
>
> - None
> ```

This is not a missing DOI as @JacksonMaxfield indicated previously in the pre-review thread (https://github.com/openjournals/joss-reviews/issues/3891#issuecomment-963817116)
