### 2021-11-30 #3904 @JacksonMaxfield
@chartgerink Zenodo now has the CDP cookiecutter archived:

Link: https://doi.org/10.5281/zenodo.5744988
DOI: 10.5281/zenodo.5744988
