### 2021-11-30 #3904 @kthyng
Hi @JacksonMaxfield! I'll be helping to wrap up this submission.

I see we have version 3.0 listed at the top of this review issue for your software, but your most recent tag is [0.0.4](https://github.com/CouncilDataProject/cookiecutter-cdp-deployment/releases/tag/v0.0.4). What version should be associated with your JOSS paper?

Also, can you manually update the Zenodo archive metadata so that the title and author list exactly match your JOSS paper?
