### 2021-11-30 #3904 @kthyng
@JacksonMaxfield Your paper looks good, but you'll need to preserve capitalization so your reference list is good. You can do this by adding {} around characters in your .bib file in which you want to preserve capitalization. One listing has R and Australia uncapitalized, but please check all carefully.
