### 2021-11-30 #3904 @JacksonMaxfield
> Hi @JacksonMaxfield! I'll be helping to wrap up this submission.

Hey @kthyng! Thanks!

> I see we have version 3.0 listed at the top of this review issue for your software, but your most recent tag is [0.0.4](https://github.com/CouncilDataProject/cookiecutter-cdp-deployment/releases/tag/v0.0.4). What version should be associated with your JOSS paper?

The paper has the correct version, I will push a v3.0.0 tag to the repo later today.

> Also, can you manually update the Zenodo archive metadata so that the title and author list exactly match your JOSS paper?

Ahhhh sorry. I used the default GitHub + Zenodo integration. From a brief look, it seems I need to add a `.zenodo.json` file and that should solve it. Will add that tonight as well.

> @JacksonMaxfield Your paper looks good, but you'll need to preserve capitalization so your reference list is good. You can do this by adding {} around characters in your .bib file in which you want to preserve capitalization. One listing has R and Australia uncapitalized, but please check all carefully.

Similarly, will do!

Thanks for the comments.
