### 2021-12-01 2c75a5f
Fix paper bib and add Zenodo JSON
