+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Trounstine2009,
-  title     = {All Politics Is Local: The Reemergence of the Study of City Politics},
+  title     = {{All Politics Is Local: The Reemergence of the Study of City Politics}},
   volume    = {7},
   doi       = {10.1017/S1537592709990892},
   url       = {https://doi.org/10.1017/S1537592709990892},
@@ -13,7 +13,7 @@
 
 @misc{cookiecutter,
   author    = {Audrey Roy Greenfeld and Daniel Roy Greenfeld and Raphael Pierzina},
-  title     = {cookiecutter},
+  title     = {{cookiecutter}},
   abstract  = {A command-line utility that creates projects from cookiecutters (project templates), e.g. creating a Python package project from a Python package project template.},
   year      = {2015},
   publisher = {GitHub},
@@ -30,13 +30,13 @@
   number    = {17},
   pages     = {411},
   author    = {Adam H. Sparks and Mark Padgham and Hugh Parsonage and Keith Pembleton},
-  title     = {bomrang: Fetch Australian Government Bureau of Meteorology Data in R},
+  title     = {{bomrang: Fetch Australian Government Bureau of Meteorology Data in R}},
   journal   = {Journal of Open Source Software}
 }
 
 @misc{courtlistener,
   author    = {{The Free Law Project}},
-  title     = {CourtListener},
+  title     = {{CourtListener}},
   abstract  = {CourtListener is a free legal research website containing millions of legal opinions from federal and state courts. With CourtListener, lawyers, journalists, academics, and the public can research an important case, stay up to date with new opinions as they are filed, or do deep analysis using our raw data.},
   year      = {2015},
   publisher = {GitHub},
@@ -46,7 +46,7 @@
 
 @misc{councilmatic,
   author    = {Mjumbe Poe and Forest Gregg},
-  title     = {Councilmatic},
+  title     = {{Councilmatic}},
   abstract  = {A subscription service for city council legislative information, started in Philadelphia.},
   year      = {2015},
   publisher = {GitHub},
@@ -63,7 +63,7 @@
 }
 
 @article{jacobi2017,
-  title     = {Justice, interrupted: The effect of gender, ideology, and seniority at Supreme Court oral arguments},
+  title     = {{Justice, interrupted: The effect of gender, ideology, and seniority at Supreme Court oral arguments}},
   author    = {Jacobi, Tonja and Schweers, Dylan},
   journal   = {Va. L. Rev.},
   volume    = {103},
@@ -74,7 +74,7 @@
 }
 
 @article{einstein2021,
-  title     = {Zoom Does Not Reduce Unequal Participation: Evidence from Public Meeting Minutes},
+  title     = {{Zoom Does Not Reduce Unequal Participation: Evidence from Public Meeting Minutes}},
   author    = {Einstein, Katherine Levine and Glick, David and Puig, Luisa Godinez and Palmer, Maxwell},
   publisher = {Housing Politics Lab},
   year      = {2021},
