+++ {{ cookiecutter.hosting_github_repo_name }}/.github/dependabot.yml
@@ -1,6 +1,13 @@
 version: 2
 updates:
   - package-ecosystem: "pip"
+    directory: "/infra"
+    schedule:
+      interval: "daily"
+    allow:
+      - dependency-name: "cdp-backend"
+
+  - package-ecosystem: "pip"
     directory: "/python"
     schedule:
       interval: "daily"
