+++ SETUP/Makefile
@@ -22,24 +22,24 @@ login: ## Run GCloud and Pulumi logins
 	gcloud auth application-default login
 
 init: ## Create google project and pulumi stack
-	gcloud projects create cdp-example-uzpjxmcp --set-as-default
-	pulumi stack init cdp-example-uzpjxmcp
+	gcloud projects create cdp-example-tnalcaxg --set-as-default
+	pulumi stack init cdp-example-tnalcaxg
 
 set-cors: ## Set cors for the bucket
-	gsutil cors set cors.json gs://cdp-example-uzpjxmcp.appspot.com/
+	gsutil cors set cors.json gs://cdp-example-tnalcaxg.appspot.com/
 
 destroy: ## Fully teardown Pulumi and GCloud
-	pulumi stack rm cdp-example-uzpjxmcp --force
-	gcloud projects delete cdp-example-uzpjxmcp
-	rm -f .keys/cdp-example-uzpjxmcp.json
+	pulumi stack rm cdp-example-tnalcaxg --force
+	gcloud projects delete cdp-example-tnalcaxg
+	rm -f .keys/cdp-example-tnalcaxg.json
 
 gen-key: ## Generate a service account JSON
 	mkdir .keys/ -p
-	gcloud iam service-accounts create cdp-example-uzpjxmcp \
+	gcloud iam service-accounts create cdp-example-tnalcaxg \
     	--description="CDP Dev Service Account for $(USER)" \
-    	--display-name="cdp-example-uzpjxmcp"
-	gcloud projects add-iam-policy-binding cdp-example-uzpjxmcp \
-    	--member="serviceAccount:cdp-example-uzpjxmcp@cdp-example-uzpjxmcp.iam.gserviceaccount.com" \
+    	--display-name="cdp-example-tnalcaxg"
+	gcloud projects add-iam-policy-binding cdp-example-tnalcaxg \
+    	--member="serviceAccount:cdp-example-tnalcaxg@cdp-example-tnalcaxg.iam.gserviceaccount.com" \
 		--role="roles/owner"
-	gcloud iam service-accounts keys create .keys/cdp-example-uzpjxmcp.json \
-  		--iam-account "cdp-example-uzpjxmcp@cdp-example-uzpjxmcp.iam.gserviceaccount.com"
+	gcloud iam service-accounts keys create .keys/cdp-example-tnalcaxg.json \
+  		--iam-account "cdp-example-tnalcaxg@cdp-example-tnalcaxg.iam.gserviceaccount.com"
