### 2021-12-02 59c3084
Bump min CDP dep versions to v3
