+++ {{ cookiecutter.hosting_github_repo_name }}/infra/requirements.txt
@@ -1,2 +1,2 @@
-cdp-backend[infrastructure]==3.0.0.dev27
+cdp-backend[infrastructure]==3.0.0
 pulumi
