+++ .github/workflows/scripts/requirements.txt
@@ -1,4 +1,4 @@
-cdp-backend[infrastructure]~=3.0.0.dev27
+cdp-backend[infrastructure]~=3.0
 cdp-scrapers~=0.3
 cookiecutter~=1.7
 requests~=2.25
