### 2021-12-02 #3904 @JacksonMaxfield
@kthyng All have now been resolved!

Want to point out that the DOI I provided previously points to Zenodo's "all versions" DOI. If that isn't what is desired, the v3.0.0 DOI is: 10.5281/zenodo.5748663
