### 2021-12-02 #3904 @kthyng
@JacksonMaxfield Ah, thank you for pointing that out with the doi. Actually I would like to use the specific DOI for the version associated with the JOSS paper so users know that for future reference as the software changes over time. With this in mind, can you update the title of *that* DOI to match your title? Looks like the author list maybe already matches?
